feat(twoslash): expose core module, independent from twoslash
