+++ packages/shikiji-twoslash/src/icons.ts
@@ -4,6 +4,6 @@ import completionIcons from './icons-completions.json'
 import tagIcons from './icons-tags.json'
 
 export type CompletionItem = NonNullable<TwoSlashReturn['queries'][0]['completions']>[0]
-export const defaultCompletionIcons: Record<CompletionItem['kind'], Element | undefined> = completionIcons as any
 
+export const defaultCompletionIcons: Record<CompletionItem['kind'], Element | undefined> = completionIcons as any
 export const defaultCustomTagIcons: Record<string, Element | undefined> = tagIcons as any
